api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiofiles>=25.1.0",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "celery[redis]>=5.6.2",
    "fastapi[standard]>=0.129.0",
    "passlib>=1.7.4",
    "pydantic-settings>=2.13.0",
    "pyjwt>=2.11.0",
    "pymupdf>=1.27.1",
    "python-multipart>=0.0.22",
    "redis>=6.4.0",
    "sqlalchemy[asyncio]>=2.0.46",
    "structlog>=25.5.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
